dateTime checked!

+++ Fiscal-Software/Fiscal-Software/Form1.cs
@@ -809,30 +809,38 @@ namespace Fiscal_Software
         private void редактиранеНаФУToolStripMenuItem_Click(object sender, EventArgs e)
         {
             // Open Fiscal Device Form for editing prefilled data
-            selectedFUDanni = int.Parse(fiscalDeviceListView.SelectedItems[0].Tag.ToString());
-            if (selectedFUDanni > 0)
+            if (fiscalDeviceListView.SelectedItems.Count > 0)
             {
-                AddFiscalDevice afd = new AddFiscalDevice(this, selectedFUDanni);
-                afd.Show();
+                selectedFUDanni = int.Parse(fiscalDeviceListView.SelectedItems[0].Tag.ToString());
+                if (selectedFUDanni > 0)
+                {
+                    AddFiscalDevice afd = new AddFiscalDevice(this, selectedFUDanni);
+                    afd.Show();
+                }
             }
+            
         }
 
         private void изтриванеНаФУToolStripMenuItem_Click(object sender, EventArgs e)
         {
             // Dialog Window to confirm deletion of the selected element
-            selectedFUDanni = int.Parse(fiscalDeviceListView.SelectedItems[0].Tag.ToString());
-            if (selectedFUDanni > 0)
+            if (fiscalDeviceListView.SelectedItems.Count > 0)
             {
-                DialogResult deleteResult = MessageBox.Show("Сигурни ли сте, че иската да изтриете това фускално у-во ?",
-                                    "Изтриване на фискално у-во",
-                            MessageBoxButtons.YesNo);
-                if (deleteResult == DialogResult.Yes)
+                selectedFUDanni = int.Parse(fiscalDeviceListView.SelectedItems[0].Tag.ToString());
+                if (selectedFUDanni > 0)
                 {
-                    int id = int.Parse(fiscalDeviceListView.SelectedItems[0].Tag.ToString());
-                    DanniFiskalnoUstroistvoCtrl.DeleteDanniFiskalnoUstroistvoById(selectedFUDanni);
-                    fiscalDeviceListView.SelectedItems[0].Remove();
+                    DialogResult deleteResult = MessageBox.Show("Сигурни ли сте, че иската да изтриете това фускално у-во ?",
+                                        "Изтриване на фискално у-во",
+                                MessageBoxButtons.YesNo);
+                    if (deleteResult == DialogResult.Yes)
+                    {
+                        int id = int.Parse(fiscalDeviceListView.SelectedItems[0].Tag.ToString());
+                        DanniFiskalnoUstroistvoCtrl.DeleteDanniFiskalnoUstroistvoById(selectedFUDanni);
+                        fiscalDeviceListView.SelectedItems[0].Remove();
+                    }
                 }
             }
+            
         }
 
         private void свидетелствоЗаРегистрацияToolStripMenuItem_Click(object sender, EventArgs e)
